requirements.txt
# Paris Municipales 2026 — Dépendances
# Instagram scraping (smart pipeline)
# Twitter/X Nitter scraping
beautifulsoup4>=4.12.0
curl_cffi>=0.6.0
playwright>=1.40.0
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0
matplotlib>=3.8.0
seaborn>=0.13.0
wordcloud>=1.9.0
networkx>=3.2.0
scikit-learn>=1.3.0
transformers>=4.36.0
torch>=2.1.0
datasets>=2.16.0
python-dateutil>=2.8.0
tqdm>=4.66.0
pytest>=7.0.0
